Implement delete product by id

+++ src/main/java/com/example/demo/service/ProductServiceImpl.java
@@ -63,7 +63,7 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProductById(Integer id) {
-        products.remove(id);
+        products.remove(id.intValue());
     }
 
     @Override
